archive/requirements.txt
# AutoReels dependencies
# Install with: pip install -r requirements.txt
# NOTE: Install torch separately first for CUDA support:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121

# Core AI / Video Generation
# torch — installed separately (see above)
diffusers>=0.30.0
transformers>=4.35.0,<5.0.0
accelerate>=0.25.0
safetensors>=0.4.0
sentencepiece>=0.2.0

# Video post-processing
moviepy>=2.0.0
Pillow>=10.0.0

# Discord bot
discord.py>=2.3.0

# Social media APIs
httpx>=0.27.0

# Config & CLI
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0

# Scheduling
apscheduler>=3.10.0

# Audio — TTS narration + music mixing
edge-tts>=7.0.0
pydub>=0.25.0
